# src/main/resources/openapi.yml
openapi: 3.0.1
info:
  title: Record Store
  description: Record Store API.
  version: 1.0.0
servers:
- url: /v2
paths:
...
  /artist/{artistId}:
    get:
      summary: get artist by id
      operationId: getArtistById
      parameters:
        - in: path
          name: artistId
          schema:
            type: integer
          required: true
          description: artist id
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtistResponseOpenApi'
    delete:
      summary: delete artist by id
      operationId: deleteArtistById
      parameters:
        - in: path
          name: artistId
          schema:
            type: integer
          required: true
          description: artist id
      responses:
        204:
          description: artist was deleted
    put:
      summary: update artist by id
      operationId: updateArtistById
      parameters:
        - in: path
          name: artistId
          schema:
            type: integer
          required: true
          description: artist id
      requestBody:
        description: artist to be updated
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArtistResponseOpenApi'
        required: true
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtistResponseOpenApi'
  /artist/new:
    post:
      summary: add new artist
      operationId: newArtist
      requestBody:
        description: artist to be added
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArtistResponseOpenApi'
        required: true
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtistResponseOpenApi'
...
components:
  schemas:
...
    StyleResponseOpenApi:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
    NationalityResponseOpenApi:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
    ArtistResponseOpenApi:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        nationality:
          $ref: '#/components/schemas/NationalityResponseOpenApi'
        style:
          $ref: '#/components/schemas/StyleResponseOpenApi'
...